fixed some issues and tested user registeration api

+++ src/services/user.service.js
@@ -5,34 +5,54 @@ import { apiError } from "../utils/apiError.js";
 import { deleteFromCloudinary } from "../utils/cloudinary.js";
 import generateAccessAndRefreshToken from "./generate.tokens.service.js";
 
-const createUser = asyncHandler(async ({ firstName, lastName, email, password, profile, phoneNumber }) => {
+const createUser = async ({
+  firstName,
+  lastName,
+  email,
+  password,
+  profile,
+  phoneNumber,
+}) => {
+  if (!firstName || !email || !password || !profile || !phoneNumber) {
+    throw new apiError(400, "All fields are required");
+  }
+  try {
+    const createdUser = await User.create({
+      fullName: {
+        firstName,
+        lastName,
+      },
+      email,
+      password,
+      profile,
+      phoneNumber,
+    })
 
-    if( !firstName || !email || !password || !profile || !phoneNumber ) {
-        throw new apiError(400, "All fields are required");
+    const user = await User.findById(createdUser._id).select("-password");
+
+    const { accessToken, refreshToken } = await generateAccessAndRefreshToken(
+      user._id,
+      User
+    );
+
+    if (!accessToken || !refreshToken) {
+      throw new apiError(400, "Token generation failed");
     }
 
-    const user = await User.create({ 
-        firstName, 
-        lastName, 
-        email, 
-        password,
-        profile,
-        phoneNumber 
-    });
-
-    const { accessToken, refreshToken } = await generateAccessAndRefreshToken(user._id, User);
-    
-    user.accessToken = accessToken;
-    user.refreshToken = refreshToken;
-
-    if( !user ) {
-        await deleteFromCloudinary(profile);
-        throw new apiError(400, "User not created");
+    if (!user) {
+      await deleteFromCloudinary(profile);
+      throw new apiError(400, "User not created");
     }
 
-    return new apiResponse(201, user);
-});
+    return {
+      user,
+      accessToken,
+      refreshToken,
+    };
+  } catch (error) {
+    await deleteFromCloudinary(profile);
+    throw new apiError(500, error?.message || "Internal server error");
+  }
+};
 
-export {
-    createUser,
-}
+export { createUser };
